refactor - core vs d3js split
